klient/machine: move mount sync event to separate file
